3.0.0 update : -add nekos.life api -help command is better -some improved commands

+++ commands/say.js
@@ -1,9 +1,11 @@
 module.exports = {
     name: 'say',
+    usage: 'pichu say',
     description: 'say things',
     category: 'misc',
     async execute(client,message,args) {
+        if (!args.join(' ')) return message.channel.send('I can\'t say nothing!')
         message.delete().then(
-        message.channel.send(args.join(' ')))
+        console.log(args.join(' ')))
     },
 }
